implemented TODOs
